+++ src/components/mod.rs
@@ -1,5 +1,6 @@
 pub mod header;
 pub mod keyboard;
 pub mod message;
+pub mod quotes;
 pub mod board;
 pub mod modal;